+++ api/redis.go
@@ -58,6 +58,17 @@ func GetTaskFromEvent(update *mesosproto.Event_Update) mesosutil.Command {
 	return mesosutil.Command{}
 }
 
+// CountRedisKey will get back the count of the redis key
+func CountRedisKey(pattern string) int {
+	keys := GetAllRedisKeys(pattern)
+	count := 0
+	for keys.Next(config.RedisCTX) {
+		count++
+	}
+	logrus.Debug("CountRedisKey: ", pattern, count)
+	return count
+}
+
 // SaveConfig store the current framework config
 func SaveConfig() {
 	data, _ := json.Marshal(config)