+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## master
+
+- ADD-11: Docker Compose Replicas
+
 ## v0.3.1
 
 - FIX-8: Set disk resource to the default value if its unset.